Separate Parser from main
